requirements.txt
Django
Pillow
python-decouple
djangorestframework
selenium
mysqlclient
textblob